requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.1
APScheduler==3.10.4
azure-ai-inference==0.1.0b2
azure-core==1.30.1
flask==3.0.2
numpy==1.26.4
python-dotenv==1.0.1
